chore: move files, remove superfluous divider
