+++ react-components/src/architecture/base/commands/DividerCommand.ts
@@ -1,0 +1,10 @@
+/*!
+ * Copyright 2024 Cognite AS
+ */
+import { RenderTargetCommand } from './RenderTargetCommand';
+
+export class DividerCommand extends RenderTargetCommand {
+  public override get isVisible(): boolean {
+    return true;
+  }
+}
