+++ react-components/src/architecture/base/concreteCommands/SettingsCommand.ts
@@ -10,7 +10,7 @@ import { SetPointShapeCommand } from './SetPointShapeCommand';
 import { PointCloudFilterCommand } from './PointCloudFilterCommand';
 import { type TranslateKey } from '../utilities/TranslateKey';
 import { type IconName } from '../utilities/IconName';
-import { PointCloudDividerCommand } from './DividerCommand';
+import { PointCloudDividerCommand } from './PointCloudDividerCommand';
 
 export class SettingsCommand extends BaseSettingsCommand {
   // ==================================================
@@ -23,7 +23,6 @@ export class SettingsCommand extends BaseSettingsCommand {
     this.add(new SetQualityCommand());
     this.add(new PointCloudDividerCommand());
     this.add(new SetPointSizeCommand());
-    this.add(new PointCloudDividerCommand());
     this.add(new SetPointColorTypeCommand());
     this.add(new SetPointShapeCommand());
     this.add(new PointCloudFilterCommand());
